add test coverage and adjust warnings
