+++ Kontent.Ai.Delivery.Abstractions/Extensions/DeliveryOptionsExtensions.cs
@@ -23,7 +23,10 @@
             o.EnableRetryPolicy = options.EnableRetryPolicy;
             o.DefaultRetryPolicyOptions = options.DefaultRetryPolicyOptions;
             o.IncludeTotalCount = options.IncludeTotalCount;
+            // See #312
+            #pragma warning disable CS0618
             o.Name = options.Name;
+            #pragma warning restore CS0618
             o.DefaultRenditionPreset = options.DefaultRenditionPreset;
         }
     }
